+++ docs/changelog.rst
@@ -9,6 +9,7 @@ Development
 - (Fill this out as you fix issues and develop your features).
 - Allow gt/gte/lt/lte/ne operators to be used with a list as value on ListField #2813
 - Switch tox to use pytest instead of legacy `python setup.py test` #2804
+- Add support for timeseries collection #2661
 
 Changes in 0.28.2
 =================